requirements.txt
fastapi
uvicorn
google-generativeai==0.7.2
supabase
python-multipart
